Add custom/special classes dynamically to a parser.
A parser instance can register additional special and custom-forced
classes by setting the  new members

  AVLTree<const String> forceCustomValueClasses;
  AVLTree<const String> specialValuedClasses;

and filling them with the class names (String).
Those will be checked in addition to the ones in the global FreeingAVLTrees:

  FreeingAVLTree<const String> defaultForceCustomValueClasses
  FreeingAVLTree<const String> defaultSpecialValuedClasses

Note that forceCustomValueClasses are only checked in case of structClass
and the names should be repeated in specialValuedClasses for them to take
full effect.
Also note that the parser will take ownership of the lists
and will dispose of them automatically.